[datalayout] Parse pointer attributes for additional address spaces
These are not yet stored, but we now use the list of pointer attributes
to set pointer_repr_size/pointer_range/pointer_align based on the
default globals address space. This is not strictly correct, but is needed
for CHERI until we use a finer-grained per-address space size/range of
pointers.
This also adds support for the 'f' pointer modifier used by CHERI LLVM to
indicate fat pointers (by ignoring it).